# frontend-next/.openapi-cache.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Learning Roadmap Generation System",
    "description": "基于 Multi-Agent 的个性化学习路线图生成系统",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tasks/{task_id}/validation/latest": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Get Latest Validation",
        "description": "获取任务的最新验证结果\n\nArgs:\n    task_id: 任务 ID\n    db: 数据库会话\n    \nReturns:\n    最新的验证记录，包含验证结果、问题列表等\n    \nRaises:\n    HTTPException: 404 - 验证记录不存在\n    \nExample:\n    ```json\n    {\n        \"id\": \"uuid\",\n        \"task_id\": \"task-123\",\n        \"roadmap_id\": \"roadmap-456\",\n        \"is_valid\": false,\n        \"overall_score\": 75.5,\n        \"issues\": {\n            \"issues\": [\n                {\n                    \"severity\": \"critical\",\n                    \"location\": \"Stage 2 > Module 1\",\n                    \"issue\": \"循环依赖检测到\",\n                    \"suggestion\": \"移除 Concept A -> B -> A 的依赖\"\n                }\n            ]\n        },\n        \"validation_round\": 1,\n        \"critical_count\": 2,\n        \"warning_count\": 5,\n        \"suggestion_count\": 3,\n        \"created_at\": \"2024-01-01T00:00:00\"\n    }\n    ```",
        "operationId": "get_latest_validation_api_v1_tasks__task_id__validation_latest_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/validation/history": {
      "get": {
        "tags": [
          "validation"
        ],
        "summary": "Get Validation History",
        "description": "获取任务的所有验证历史记录\n\nArgs:\n    task_id: 任务 ID\n    db: 数据库会话\n    \nReturns:\n    验证记录列表（按轮次降序排列）\n    \nExample:\n    ```json\n    [\n        {\n            \"id\": \"uuid-1\",\n            \"task_id\": \"task-123\",\n            \"validation_round\": 2,\n            \"is_valid\": true,\n            \"overall_score\": 92.0,\n            \"created_at\": \"2024-01-01T01:00:00\"\n        },\n        {\n            \"id\": \"uuid-2\",\n            \"task_id\": \"task-123\",\n            \"validation_round\": 1,\n            \"is_valid\": false,\n            \"overall_score\": 75.5,\n            \"created_at\": \"2024-01-01T00:00:00\"\n        }\n    ]\n    ```",
        "operationId": "get_validation_history_api_v1_tasks__task_id__validation_history_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/jwt/login"
          }
        }
      }
    }
  }
}